# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.5.0",
    "title": "Kubero",
    "description": "Kubero is a web-based tool deploy applications on a Kubernetes clusters. It provides a simple and intuitive interface to manage your clusters, applications, and pipelines."
  },
  "servers": [
    {
      "url": "http://localhost:2000/api"
    }
  ],
  "paths": {
...
    "/repo/{repoprovider}/list": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "Get a list of available repositories",
        "description": "",
        "parameters": [
          {
            "name": "repoprovider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/repo/{repoprovider}/connect": {
      "post": {
        "tags": [
          "Repositories"
        ],
        "summary": "Connect a repository to a pipeline",
        "description": "",
        "parameters": [
          {
            "name": "repoprovider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "gitrepo": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/repo/{repoprovider}/{gitrepob64}/branches/list": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "Get a list of available branches",
        "description": "",
        "parameters": [
          {
            "name": "repoprovider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "github",
                "gitlab",
                "bitbucket",
                "gitea",
                "gogs"
              ]
            },
            "description": "Repository provider"
          },
          {
            "name": "gitrepob64",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Base64 encoded git repository url"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}